refactor: extract terminal emulator behind a backend trait
The grid was produced directly by an alacritty-specific `Emu` struct, so
supporting another emulator meant editing every consumer. Split the seam:

- `terminal/cell.rs` holds the neutral grid vocabulary (`EmuCell`, `Color`,
  `rows_to_strings`) with no emulator dependency
- `terminal/emu.rs` defines the `Emulator` trait, the whole contract between
  the PTY and the grid
- `terminal/alacritty.rs` (moved from `emu.rs`) implements it

Command/exit/cwd tracking moves out of the emulator into `TermState`. It is
derived from the raw PTY byte stream and never touched the alacritty grid, so
keeping it outside the trait means every backend reports identical shell
integration by construction instead of reimplementing it. `integration.rs` is
unchanged.

Consumers already spoke `EmuCell`/`Color` rather than alacritty types, so this
is import churn for them. No behavior change.
