+++ src/terminal/mod.rs
@@ -1,3 +1,5 @@
+pub mod alacritty;
+pub mod cell;
 pub mod emu;
 pub mod integration;
 pub mod locator;
